move `useCurrentLanguage` to a proper file
