+++ src/components/TimeTable/LectureCard.tsx
@@ -20,7 +20,7 @@ import Animated, {
   useSharedValue,
   withTiming,
 } from "react-native-reanimated"
-import { useCurrentLanguage } from "utils/articles"
+import { useCurrentLanguage } from "utils/language"
 
 export interface LectureCardProps {
   lecture: Event
