+++ src/pages/news/ArticleDetails.tsx
@@ -1,11 +1,8 @@
 import { MainStackScreen } from "navigation/NavigationTypes"
 import { ArticleDetailsWrapper } from "components/Home/News/ArticleDetailsWrapper"
 import React, { useCallback, useRef, useState } from "react"
-import {
-  extractImageLinks,
-  useCurrentLanguage,
-  getArticleParams,
-} from "utils/articles"
+import { extractImageLinks, getArticleParams } from "utils/articles"
+import { useCurrentLanguage } from "utils/language"
 import { ModalSlider } from "components/Home/News/ModalSlider/ModalSlider"
 import { MemoizedMarkdown } from "components/Home/News/ModalSlider/MemoizedMarkdown"
 import { Image } from "expo-image"
